app/src/main/java/com/padma/testingassigmentandroid/data/LoginDataSource.kt
package com.padma.testingassigmentandroid.data

import com.padma.testingassigmentandroid.data.model.LoggedInUser
import java.io.IOException
import java.lang.Exception
import java.util.regex.Pattern

/**
 * Class that handles authentication w/ login credentials and retrieves user information.
 */
class LoginDataSource {

    fun login(email: String, password: String): Result<LoggedInUser> {

        return try {
            val isValid = isValidEmailPattern(email);
            if(!isValid){
                throw Exception("This is not an email")
            }
            // TODO: handle loggedInUser authentication
            val fakeUser = LoggedInUser(java.util.UUID.randomUUID().toString(), "Dear")
            Result.Success(fakeUser)
        } catch (e: Throwable) {
            Result.Error(IOException("Error logging in", e))
        }
    }

    fun logout() {
        // TODO: revoke authentication
    }

    fun isValidEmailPattern(email: String):Boolean {
         return Pattern.compile(
            "^(([\\w-]+\\.)+[\\w-]+|([a-zA-Z]|[\\w-]{2,}))@"
                    + "((([0-1]?[0-9]{1,2}|25[0-5]|2[0-4][0-9])\\.([0-1]?"
                    + "[0-9]{1,2}|25[0-5]|2[0-4][0-9])\\."
                    + "([0-1]?[0-9]{1,2}|25[0-5]|2[0-4][0-9])\\.([0-1]?"
                    + "[0-9]{1,2}|25[0-5]|2[0-4][0-9]))|"
                    + "([a-zA-Z]+[\\w-]+\\.)+[a-zA-Z]{2,4})$"
        ).matcher(email).matches()
    }

    fun additionOfTwoNumbers(number1:Int, number2: Int):Int {
        return number1+ number2;
    }
}